pyproject.toml
[project]
name = "healher-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtplib>=5.1.0",
    "bcrypt>=4.0.0",
    "certifi>=2026.2.25",
    "python-dotenv>=1.0.0",
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-core>=1.2.23",
    "langchain-nvidia-ai-endpoints>=1.2.1",
    "langchain-pinecone>=0.2.13",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-redis>=0.4.0",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4.1",
    "pinecone>=7.3.0",
    "pydantic[email]>=2.12.5",
    "requests>=2.33.0",
    "uvicorn>=0.42.0",
]
